requirements.txt
asgiref==3.8.1
Django==5.1.5
python-decouple==3.8
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2024.2
wheel==0.45.1
